+++ running/ui/pageeditrecord.py
@@ -3,8 +3,8 @@ from tkinter import ttk, messagebox
 import datetime
 
 from running import SessionRunning, DatabaseManagerRunning
-from ui import Page
 from .pageaddrecord import PageAddRecord
+
 
 class PageEditRecord(PageAddRecord):