+++ ui/widgets/__init__.py
@@ -1,1 +1,4 @@
-from .framescrollable import FrameScrollable
+from .page import Page
+from .framescrollablex import FrameScrollableX
+from .framescrollabley import FrameScrollableY
+from .framescrollablexy import FrameScrollableXY